pyproject.toml
[project]
name = "jsontopia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "jsonschema>=4.23.0",
    "langchain-openai>=0.3.13",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
